# perf(crypto/merkle, crypto/tmhash): simplify+optimize SHA256 hashing of multiple byteslices