--- a/.changelog/unreleased/improvements/1921-crypto-merkle-innerHash.md
+++ b/.changelog/unreleased/improvements/1921-crypto-merkle-innerHash.md
@@ -1,0 +1,1 @@
+- `[crypto/merkle]` faster calculation of hashes ([#1921](https://github.com/cometbft/cometbft/pull/1921))