--- a/crypto/merkle/hash.go
+++ b/crypto/merkle/hash.go
@@ -32,11 +32,7 @@ func leafHashOpt(s hash.Hash, leaf []byte) []byte {
 
 // returns tmhash(0x01 || left || right).
 func innerHash(left []byte, right []byte) []byte {
-	data := make([]byte, len(innerPrefix)+len(left)+len(right))
-	n := copy(data, innerPrefix)
-	n += copy(data[n:], left)
-	copy(data[n:], right)
-	return tmhash.Sum(data)
+	return tmhash.SumMany(innerPrefix, left, right)
 }
 
 func innerHashOpt(s hash.Hash, left []byte, right []byte) []byte {