@odeke-em 2024-01-03
>LGTM, thanks for reducing allocation overhed! Though the changelog comment of "much more efficient" seems a bit odd. Its a non-existent time speedup, and a 10% allocation reduction!

Thanks for the reivew @ValarDragon! I don't follow what you mean by "non-existent time speedup and 10% allocation reduction" precluding efficiency -- innerHash is called a whole lot inside crypto/merkle which is a critical package, in fact even if it were a 5% allocation reduction I'd call it that because the API change means no wastage of CPU cycles but also the runtime doing less work. Also we've had quite the rhetoric in the Cosmos community of "optimizations must be huge updates or re-engineer the whole thing" but sadly that leads to a ton of inefficient code patterns that linger for 7+ years, moreover very few people (except yourself and Osmosis) actually profile code but even then folks aren't looking at profiles intently hence massive bloats. 

Thanks for the code reviews @ValarDragon and @melekes!