@melekes 2024-01-05
@odeke-em could you please fix the build error:

```
# github.com/cometbft/cometbft/crypto/merkle [github.com/cometbft/cometbft/crypto/merkle.test]
Error: crypto/merkle/bench_test.go:4:2: "bytes" imported and not used
```